requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
audioread==3.0.1
cachetools==5.5.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
fastapi==0.114.1
ffmpy==0.4.0
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
gradio==4.44.0
gradio_client==1.3.0
greenlet==3.1.0
grpcio==1.66.1
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.2.5
langchain-core==0.2.39
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.1
langsmith==0.1.118
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.3
numba==0.61.2
numpy==2.2.5
orjson==3.10.7
packaging==25.0
pandas==2.2.2
pillow==10.4.0
pinecone==5.1.0
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
pooch==1.8.2
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rsa==4.9
ruff==0.6.4
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
setuptools==74.1.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.34
starlette==0.38.5
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing_extensions==4.13.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.30.6
websockets==12.0
wheel==0.44.0
yarl==1.11.1
